# contracts/http/control-plane.openapi.yaml
openapi: 3.1.0
info:
  title: Back Pressure Control Plane API
  version: 0.1.0-phase1a
  description: Phase 1A contract freeze for the Spring Boot control-plane service.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/scenario/current:
    put:
      operationId: putCurrentScenario
      summary: Create or replace the current in-memory scenario configuration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/scenario-config.schema.json
      responses:
        '200':
          description: Accepted scenario configuration.
          content:
            application/json:
              schema:
                $ref: ../schemas/scenario-config.schema.json
...